Check amount of overlap Semmed gold standard has with Hetionet
We would like to run Deepwalk on Hetionet to get an idea of the
performance. However, since the network in Hetionet is different
from SemmedDB, we are concerned with whether we will be able to
use the SemmedDB gold standard for evaluation.

This notebook checks to see how much of the SemmedDB gold standard
can be found in Hetionet.