Add edge subsampler
Create the notebook which subsamples edges to create new networks
so that we can test whether specific edge types are more crucial
to the predictive framework.

Current subsampler tries to keep all nodes in the gold standard.
If the removed edges remove nodes which appear in the gold standard,
then the edges are resubsampled. If after multiple attempts the
gold standard nodes cannot be preserved, then the edge type is
skipped.